+++ skopt/utils.py
@@ -264,15 +264,20 @@ def cook_estimator(base_estimator, space=None, **kwargs):
     """
     Cook a default estimator.
 
+    For the special base_estimator called "DUMMY" the return value is None.
+    This corresponds to sampling points at random, hence there is no need
+    for an estimator.
+
     Parameters
     ----------
-    * `base_estimator` ["GP", "RF", "ET", "GBRT" or sklearn regressor, default="GP"]:
+    * `base_estimator` ["GP", "RF", "ET", "GBRT", "DUMMY"
+                        or sklearn regressor, default="GP"]:
         Should inherit from `sklearn.base.RegressorMixin`.
         In addition the `predict` method should have an optional `return_std`
         argument, which returns `std(Y | x)`` along with `E[Y | x]`.
-        If base_estimator is one of ["GP", "RF", "ET", "GBRT"], a default
-        surrogate model of the corresponding type is used corresponding to what
-        is used in the minimize functions.
+        If base_estimator is one of ["GP", "RF", "ET", "GBRT", "DUMMY"], a
+        surrogate model corresponding to the relevant `X_minimize` function
+        is created.
 
     * `space` [Space instance]:
         Has to be provided if the base_estimator is a gaussian process.
@@ -286,9 +291,10 @@ def cook_estimator(base_estimator, space=None, **kwargs):
         is_cat = space.is_categorical
     if isinstance(base_estimator, str):
         base_estimator = base_estimator.upper()
-        if base_estimator not in ["GP", "ET", "RF", "GBRT"]:
+        if base_estimator not in ["GP", "ET", "RF", "GBRT", "DUMMY"]:
             raise ValueError("Valid strings for the base_estimator parameter "
-                             " are: 'RF', 'ET' or 'GP' not %s" % base_estimator)
+                             " are: 'RF', 'ET', 'GP', 'GBRT' or 'DUMMY' not "
+                             "%s." % base_estimator)
     elif not is_regressor(base_estimator):
         raise ValueError("base_estimator has to be a regressor.")
 
@@ -315,5 +321,103 @@ def cook_estimator(base_estimator, space=None, **kwargs):
         gbrt = GradientBoostingRegressor(n_estimators=30, loss="quantile")
         base_estimator = GradientBoostingQuantileRegressor(base_estimator=gbrt)
 
+    elif base_estimator == "DUMMY":
+        return None
+
     base_estimator.set_params(**kwargs)
     return base_estimator
+
+
+def dimensions_aslist(search_space):
+    """Convert a dict representation of a search space into a list of
+    dimensions, ordered by sorted(search_space.keys()).
+
+    Parameters
+    ----------
+    search_space : dict
+        Represents search space. The keys are dimension names (strings)
+        and values are instances of classes that inherit from the class
+        skopt.space.Dimension (Real, Integer or Categorical)
+        Example:
+            {'name1': Real(0,1), 'name2': Integer(2,4), 'name3': Real(-1,1)}
+
+    Returns
+    -------
+    params_space_list: list of skopt.space.Dimension instances.
+        Example output with example inputs:
+            [Real(0,1), Integer(2,4), Real(-1,1)]
+    """
+    params_space_list = [
+        search_space[k] for k in sorted(search_space.keys())
+    ]
+    return params_space_list
+
+
+def point_asdict(search_space, point_as_list):
+    """Convert the list representation of a point from a search space
+    to the dictionary representation, where keys are dimension names
+    and values are corresponding to the values of dimensions in the list.
+
+    Counterpart to parameters_aslist.
+
+    Parameters
+    ----------
+    search_space : dict
+        Represents search space. The keys are dimension names (strings)
+        and values are instances of classes that inherit from the class
+        skopt.space.Dimension (Real, Integer or Categorical)
+        Example:
+            {'name1': Real(0,1), 'name2': Integer(2,4), 'name3': Real(-1,1)}
+
+    point_as_list : list
+        list with parameter values.The order of parameters in the list
+        is given by sorted(params_space.keys()).
+        Example:
+            [0.66, 3, -0.15]
+
+    Returns
+    -------
+    params_dict: dictionary with parameter names as keys to which
+        corresponding parameter values are assigned.
+        Example output with inputs:
+            {'name1': 0.66, 'name2': 3, 'name3': -0.15}
+    """
+    params_dict = {
+        k: v for k, v in zip(sorted(search_space.keys()), point_as_list)
+    }
+    return params_dict
+
+
+def point_aslist(search_space, point_as_dict):
+    """Convert a dictionary representation of a point from a search space to
+    the list representation. The list of values is created from the values of
+    the dictionary, sorted by the names of dimensions used as keys.
+
+    Counterpart to parameters_asdict.
+
+    Parameters
+    ----------
+    search_space : dict
+        Represents search space. The keys are dimension names (strings)
+        and values are instances of classes that inherit from the class
+        skopt.space.Dimension (Real, Integer or Categorical)
+        Example:
+            {'name1': Real(0,1), 'name2': Integer(2,4), 'name3': Real(-1,1)}
+
+    point_as_dict : dict
+        dict with parameter names as keys to which corresponding
+        parameter values are assigned.
+        Example:
+            {'name1': 0.66, 'name2': 3, 'name3': -0.15}
+
+    Returns
+    -------
+    point_as_list: list with point values.The order of
+        parameters in the list is given by sorted(params_space.keys()).
+        Example output with example inputs:
+            [0.66, 3, -0.15]
+    """
+    point_as_list = [
+        point_as_dict[k] for k in sorted(search_space.keys())
+    ]
+    return point_as_list